Improve copyProperties() performance.
For some reason `Object.values()` is rather expensive in Jint, the JS engine used by PokeAByte. The new implementation avoids one `values()` call completey and uses `Object.keys` in place of the other, because that does not seem to have quite the same penalty.

I also slightly improved how we create the source path.

This this mapper change reduces the CPU load for the "pokemon_emerald" mapper on my working branch of PokeAByte from ~11% to ~7.2%. And for "pokemon_yellow" it goes from ~30% down to ~17%.

There is more potential for optimizations, but none quite as impactful as this change.